Remove Tracking.js and move contents to nav.js, add comments
