+++ atd-vzv/src/views/nav/SideDrawerMobileNav.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { A, usePath } from "hookrouter";
 import { Button, Nav, NavItem, NavLink } from "reactstrap";
-import { navConfig } from "../../constants/nav";
-import { trackPageEvent } from "../../Components/Tracking/Tracking";
+import { navConfig, trackPageEvent } from "../../constants/nav";
 import { responsive } from "../../constants/responsive";
 import { colors } from "../../constants/colors";
 import styled from "styled-components";
